+++ source/rasterizer/renderer.cpp
@@ -7,8 +7,8 @@ namespace Rasterizer
 	void Clear(const FImageView& ColorBuffer, const FVector4f& Color)
 	{
 		FColor4UB* Pixels = ColorBuffer.Pixels;
-		float Size = ColorBuffer.width * ColorBuffer.height;
-		for (int i = 0; i < Size; i++)
+		uint32_t Size = ColorBuffer.width * ColorBuffer.height;
+		for (uint32_t i = 0; i < Size; i++)
 		{
 			Pixels[i] = Vector4fToColor4UB(Color);
 		}
